chore(webpack): refactor webpack for best performance
