--- client/src/assets/font-awesome/less/fixed-width.less
@@ -1,6 +1,0 @@
-// Fixed Width Icons
-// -------------------------
-.@{fa-css-prefix}-fw {
-  width: (18em / 14);
-  text-align: center;
-}
