--- client/src/assets/font-awesome/less/screen-reader.less
@@ -1,5 +1,0 @@
-// Screen Readers
-// -------------------------
-
-.sr-only { .sr-only(); }
-.sr-only-focusable { .sr-only-focusable(); }
