--- client/src/assets/font-awesome/scss/_fixed-width.scss
@@ -1,6 +1,0 @@
-// Fixed Width Icons
-// -------------------------
-.#{$fa-css-prefix}-fw {
-  width: (18em / 14);
-  text-align: center;
-}
