--- client/src/assets/font-awesome/scss/_screen-reader.scss
@@ -1,5 +1,0 @@
-// Screen Readers
-// -------------------------
-
-.sr-only { @include sr-only(); }
-.sr-only-focusable { @include sr-only-focusable(); }
